Support building against system libraries
Problem
-------

**Given**:

A sandbox environment is an environment where dependencies
are all provided but access to the internet is not. Also take as
a given that the build for this project happens in an environment
that meets that description.

**When**:

Configuring this project with any of the provided approaches. For
instance:

```
cmake -B build -S .
```

**Then**:

The configuration step hangs or fails, depending on the nature of
the sandboxed environment.

Solution
--------

Introduce a `BEMAN_EXEMPLAR_FETCH_GOOGLETEST` option to allow the
user to disable actions to download files from the internet without
disabling the building and running of tests. Those dependencies will
be provided correctly such that `find_package` calls will complete
successfully.

This workflow will provide minimal support for all known existing
packaging approaches, though `BEMAN_EXEMPLAR_FETCH_GOOGLETEST`
option will need to be explicitly set to a falsey value.

Future commits can revisit how to balance `FetchContent` and
`find_package` approaches to configuring CMake projects.