Support system libs
Problem
-------

**Given**:

A sandbox environment is an environment where dependencies
are all provided but access to the internet is not. Also take as
a given that the build for this project happens in an environment
that meets that description.

**When**:

Configuring this project with any of the provided approaches. For
instance:

```
cmake -B build -S .
```

**Then**:

The configuration step hangs or fails, depending on the nature of
the sandboxed environment.

Solution
--------

Alter `FetchContent_*` calls in this project to use `GTest` as the
project name and nothing else. This is important to allow the
`FETCHCONENT_TRY_FIND_PACKAGE_MODE` feature to work as designed.
This allows the human or tool building this project to explicitly
disable operations involving calls to `git` or other off-box I/O.

The documentation in the README was also updated to include
workflow tweaks for building against packaged versions of either
existing GoogleTest source code or prebuilt GoogleTest libraries.