Inject FetchContent logic
Problem
-------

**Given**:

A sandbox environment is an environment where dependencies
are all provided but access to the internet is not. Also take as
a given that the build for this project happens in an environment
that meets that description.

**When**:

Configuring this project with any of the provided approaches. For
instance:

```
cmake -B build -S .
```

**Then**:

The configuration step hangs or fails, depending on the nature of
the sandboxed environment.

Solution
--------

Move to a `find_package` first implementation for the
`CMakeLists.txt` of this project.

Using `FetchContent` is still trivial, using the new functionality
in `cmake/use-fetch-content.cmake` as documented in the updates to
`README.md` included in this commit.